+++ src/game/game.js
@@ -16,7 +16,7 @@ export class Game {
         this.placeBetButton = document.getElementById(`place-${this.bet.name}`);
         this.hitButton = document.getElementById(this.hit.name);
         this.standButton = document.getElementById(this.stand.name);
-        this.howToPlayButton = document.getElementById('how-to-play');
+        this.howToPlayOpenButton = document.getElementById('open-how-to-play');
         this.howToPlayDialog = document.getElementById('how-to-play-dialog');
         this.howToPlayCloseButton = document.getElementById('close-how-to-play');
         this.DEALER_DELAY_MS = 4e2;
@@ -124,9 +124,8 @@ export class Game {
         this.betListener();
     }
     howToPlayListener() {
-        // TODO show rules and controls
-        this.howToPlayButton.addEventListener('click', () => {
-            this.ripple(this.howToPlayButton);
+        this.howToPlayOpenButton.addEventListener('click', () => {
+            this.ripple(this.howToPlayOpenButton);
             this.howToPlayDialog.showModal();
         });
         this.howToPlayCloseButton.addEventListener('click', () => this.howToPlayDialog.close());
@@ -162,12 +161,17 @@ export class Game {
         this.placeBetButton.addEventListener('blur', () => this.placeBetButton.focus());
         this.placeBetButton.addEventListener('focus', () => this.betInput.step = this.betInputMin);
         this.placeBetButton.addEventListener('wheel', (event) => (event.deltaY > 0) ? this.betInput.stepDown() : this.betInput.stepUp());
-        this.placeBetButton.addEventListener('click', () => {
+        this.placeBetButton.addEventListener('click', () => __awaiter(this, void 0, void 0, function* () {
             this.hitEnterKeyListener();
             this.bet();
             this.updateButtons(false, false, [this.placeBetButton]);
-            this.updateButtons(true, true, [this.hitButton, this.standButton]);
-        });
+            const actionButtons = [this.hitButton, this.standButton];
+            if (this.player.hasBlackjack()) {
+                this.updateButtons(true, false, actionButtons);
+                return yield new Promise(resolve => setTimeout(resolve, 3e3)).then(() => this.stand());
+            }
+            this.updateButtons(true, true, actionButtons);
+        }));
     }
     /**
      * Allow to use the 'Enter' key to hit after placing a bet