+++ src/role/role.js
@@ -21,23 +21,29 @@ export class Role {
     get score() {
         return this.hand.map(card => card.value).reduce((prev, curr) => prev + curr, 0);
     }
+    get blackjack() {
+        return (this.hand.length === 2) && (Rules.BLACKJACK === this.score);
+    }
     addCard() {
+        var _a;
         const suit = this.getRandomEnum(Suit);
         const rank = this.getRandomEnum(Rank);
         if (Game.cardsDealt.some(card => (card.suit === suit) && (card.rank === rank))) {
             return this.addCard();
         }
         const card = new Card(suit, rank);
-        if ((Rank.ACE === card.rank) && ((this.score + card.value) > Rules.BLACKJACK)) {
-            card.dictionary.set(Rank.ACE, 1);
-        }
         Game.cardsDealt.push(card);
         this.hand.push(card);
+        if ((this.score > Rules.BLACKJACK)) {
+            (_a = this.hand.find(card => (card.rank === Rank.ACE) && (card.dictionary.get(Rank.ACE) === 11))) === null || _a === void 0 ? void 0 : _a.dictionary.set(Rank.ACE, 1);
+        }
         this.writeCard(card);
         this.writeScore();
     }
     writeScore(score) {
+        this.scoreBox.classList.add('refresh-value');
         this.scoreBox.innerText = score !== null && score !== void 0 ? score : this.score.toString();
+        setTimeout(() => this.scoreBox.classList.remove('refresh-value'), 5e2);
     }
     writeCard(card, isHidden) {
         var _a, _b;
@@ -52,9 +58,6 @@ export class Role {
         const keys = Object.keys(enumeration).filter(key => Number.isNaN(Number.parseInt(key)));
         const enumKey = keys[Math.floor(Math.random() * keys.length)];
         return enumeration[enumKey];
-    }
-    hasBlackjack() {
-        return (this.hand.length === 2) && (Rules.BLACKJACK === this.score);
     }
     clearHand() {
         return __awaiter(this, void 0, void 0, function* () {