+++ src/role/role.ts
@@ -18,6 +18,10 @@ export abstract class Role {
         return this.hand.map(card => card.value).reduce((prev, curr) => prev + curr, 0);
     }
 
+    get blackjack(): boolean {
+        return (this.hand.length === 2) && (Rules.BLACKJACK === this.score);
+    }
+
     addCard(): void {
         const suit = this.getRandomEnum(Suit);
         const rank = this.getRandomEnum(Rank);
@@ -28,18 +32,21 @@ export abstract class Role {
 
         const card = new Card(suit, rank);
 
-        if ((Rank.ACE === card.rank) && ((this.score + card.value) > Rules.BLACKJACK)) {
-            card.dictionary.set(Rank.ACE, 1);
-        }
-
         Game.cardsDealt.push(card);
         this.hand.push(card);
+
+        if ((this.score > Rules.BLACKJACK)) {
+            this.hand.find(card => (card.rank === Rank.ACE) && (card.dictionary.get(Rank.ACE) === 11))?.dictionary.set(Rank.ACE, 1);
+        }
+
         this.writeCard(card);
         this.writeScore();
     }
 
     protected writeScore(score?: string) {
+        this.scoreBox.classList.add('refresh-value');
         this.scoreBox.innerText = score ?? this.score.toString();
+        setTimeout(() => this.scoreBox.classList.remove('refresh-value'), 5e2);
     }
 
     protected writeCard(card: Card, isHidden?: boolean): void {
@@ -58,10 +65,6 @@ export abstract class Role {
         const enumKey = keys[Math.floor(Math.random() * keys.length)];
 
         return enumeration[enumKey];
-    }
-
-    hasBlackjack(): boolean {
-        return (this.hand.length === 2) && (Rules.BLACKJACK === this.score);
     }
 
     async clearHand(): Promise<string> {